updated data sets and some classes, added repos and services

+++ src/main/java/hr/tvz/quizzard/repository/UserEntityRepository.java
@@ -3,5 +3,5 @@ package hr.tvz.quizzard.repository;
 import hr.tvz.quizzard.model.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserEntityRepository extends JpaRepository<UserEntity, Long> {
+public interface UserEntityRepository extends JpaRepository<UserEntity, Integer> {
 }
